microservices/query/requirements.txt
fastapi
uvicorn
google-cloud-retail
google-auth
